+++ server/src/api/routes/competitions.ts
