+++ server/src/api/routes/names.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import * as controller from '@controllers/names';
-import * as validator from './name.validator';
+import * as validator from '../modules/names/name.validator';
 
 const api = Router();
